refactor! Refactor CLI tests
